scripts/requirements.txt
boto3>=1.20.0,<2.0.0
supabase>=1.0.0,<2.0.0
Pillow>=9.0.0,<11.0.0
requests>=2.25.0,<3.0.0 